export in 8 (unsigned), 16 (signed), 24 (packed) and 32 (floating point) bit depth

+++ include/files/export.h
@@ -6,6 +6,6 @@
 #ifndef EXPORT_H
 #define EXPORT_H
 
-int makeWavHeader(FILE *file, uint32_t sampleRate, uint16_t bitDepth, uint32_t sampleCount);
+int makeWavHeader(FILE *file, uint32_t sampleRate, uint16_t bitDepth, uint64_t sampleCount);
 
 #endif
